+++ lib/functions/_types.ts
@@ -1,3 +1,0 @@
-declare const brand: unique symbol;
-
-export type Brand<T, TBrand extends string> = T & { [brand]: TBrand };