+++ lib/functions/vGuid.ts
@@ -1,8 +1,6 @@
-import {type Brand} from './_types';
+import {type GUID} from '../types';
 
 const RGX = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-
-export type GUID = Brand<string, 'GUID'>;
 
 /**
  * Validate that a provided value is a valid guid according to rfc 4122