+++ lib/functions/vSysIPv4.ts
@@ -1,8 +1,6 @@
-import {type Brand} from './_types';
+import {type IP_V4} from '../types';
 
 export const RGX = /^(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/; /* eslint-disable-line max-len */
-
-export type IP_V4 = Brand<string, 'IP_V4'>;
 
 /**
  * Validate that a provided value is a valid IPv4 address